[2MGFX] Split ShaderInfo.
ShaderInfo keeps only the data that are related to the parsed node.
Everything else -including ShaderInfo- are moved to ShaderResult.

The idea is to seperate the parser from everything else in order to swap
it with the HLSL parser by @TGJones.
As a next step we can move everything related to the parser and tree
nodes to another folder and namespace (TwoMGFX.TPGParser).
